perf(api): lazy facade for the distilled Cloudflare SDK — full-graph eval 84→78ms
CloudflareApi.ts becomes a dynamic-import facade; the distilled SDK wrapper
moves to CloudflareApiImpl.ts (loaded on first Cloudflare-integration call).
Drops ~2.4MB of SDK source (~250KB minified) from the request-path closure.

Scorecard: eval 119/84 -> 127/78 ms, heap 25.7/40.7 -> 25.7/39.9 MB.
